scripts/requirements.txt
#(python)?\s*3\.(10|11)\..*
wcwidth
pytest
pytest-asyncio
sphinx-rtd-theme
sphinxcontrib-mermaid
sphinx-markdown-builder
myst-parser
Pygments
pycryptodome
PyYAML
paramiko
reedsolo
